+++ guniflask/orm/__init__.py
@@ -1,5 +1,9 @@
 # coding=utf-8
 
-from .sql_alchemy_wrapper import *
+from .base_model import *
+from .model_utils import *
+from .sqlalchemy_wrapper import *
 
-__all__ = sql_alchemy_wrapper.__all__
+__all__ = (base_model.__all__ +
+           model_utils.__all__ +
+           sqlalchemy_wrapper.__all__)